requirements.txt
requests
redis
python-dotenv
Flask
flask-cors
PyJWT